autoload psr-4; add cs-fixer; add library for .env management; add js and css libraries
