+++ .env.example
@@ -1,0 +1,4 @@
+GH_REPOSITORIES=repsitory1|repository2
+GH_ACCOUNT=account
+GH_CLIENT_ID=client_id
+GH_CLIENT_SECRET=client_secret
