+++ .gitignore
@@ -1,2 +1,5 @@
 .vscode
 vendor
+.env
+.idea
+.php_cs.cache
