+++ .php_cs
@@ -1,0 +1,14 @@
+<?php
+
+$finder = PhpCsFixer\Finder::create()
+    ->exclude('vendor')
+    ->in(__DIR__)
+
+;
+
+return PhpCsFixer\Config::create()
+    ->setRules([
+        'array_syntax' => ['syntax' => 'short'],
+        'declare_strict_types' => true
+    ])
+    ->setFinder($finder);
