+++ composer.json
@@ -1,18 +1,20 @@
 {
-	"name": "sk/centra-backend-task",
-	"description": "Kanban board for github",
-	"version": "0.0.1",
-	"autoload": {
-		"classmap": [
-			"src/classes"
-		]
-	},
-	"require": {
-		"mustache/mustache": "^2.9",
-		"vierbergenlars/php-semver": "^3.0",
-		"knplabs/github-api": "^1.5",
-		"michelf/php-markdown": "^1.6",
-		"phpunit/phpunit": "5.0.*"
-	},
-	"minimum-stability": "stable"
+  "autoload": {
+    "psr-4": {
+      "App\\": "src/"
+    }
+  },
+  "require": {
+    "mustache/mustache": "^2.9",
+    "vierbergenlars/php-semver": "^3.0",
+    "knplabs/github-api": "^1.5",
+    "michelf/php-markdown": "^1.6",
+    "vlucas/phpdotenv": "^4.1",
+    "symfony/event-dispatcher": "^4.3"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "5.0.*",
+    "friendsofphp/php-cs-fixer": "^2.16"
+  },
+  "minimum-stability": "stable"
 }
