+++ notes.md
@@ -1,12 +1,12 @@
 # Notes
 
 ## Plan
-### 1. Create repo in Gitlab or Github
-### 2. Prepare docker environment
-### 3. Run original code 
-### 4. Refactor
-### 5. Run some server 
-
+1. Create repo in Gitlab or Github
+2. Prepare docker environment
+3. Run original code
+    - add .env
+4. Refactor
+5. Run some server - digital
 
 ## Q&A
 ### What is github kanban board?
