+++ src/views/index.mustache
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
+<head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -8,33 +8,37 @@
     <title>Kanban Board</title>
 
     <!-- Bootstrap -->
-    <link href="css/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+          integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
     <style>
         [class*="col-"] {
             padding-top: 0px;
             padding-bottom: 0px;
         }
+
         .nopadding {
-            padding:0px;
+            padding: 0px;
             margin: 4px;
         }
+
         .progress {
             margin-bottom: 0 !important;
         }
+
         .list-group-item-warning #state:after {
-            font-family:'Glyphicons Halflings';
-            content:"\e073";
+            font-family: 'Glyphicons Halflings';
+            content: "\e073";
         }
     </style>
     <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
     <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
     <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-      <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
+    <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
+    <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
     <![endif]-->
-  </head>
-  <body>
-  <div class="container-fluid">
+</head>
+<body>
+<div class="container-fluid">
     {{#milestones}}
         <div class="panel panel-primary">
             <div class="panel-heading">
@@ -61,7 +65,8 @@
                     <div class="col-sm-4">
                         <div class="list-group">
                             {{#active}}
-                                <a href="{{url}}" class="list-group-item {{#paused}}list-group-item-warning{{/paused}} {{^paused}}list-group-item-info{{/paused}}">
+                                <a href="{{url}}"
+                                   class="list-group-item {{#paused}}list-group-item-warning{{/paused}} {{^paused}}list-group-item-info{{/paused}}">
                                     <img class="img-rounded" src="{{assignee}}" width="16" height="16" border="0"/>&nbsp;
                                     {{#paused}}<em><small>{{/paused}}{{title}}{{#paused}}</small></em>{{/paused}}
                                     <span id="state" class="glyphicon"></span>
@@ -86,16 +91,23 @@
             <div class="panel-footer small">
                 {{#progress}}
                     <div class="progress">
-                        <div class="progress-bar" role="progressbar" aria-valuenow="{{percent}}" aria-valuemin="0" aria-valuemax="100" style="min-width: 2em; width: {{percent}}%;">{{percent}}%</div>
+                        <div class="progress-bar" role="progressbar" aria-valuenow="{{percent}}" aria-valuemin="0"
+                             aria-valuemax="100" style="min-width: 2em; width: {{percent}}%;">{{percent}}%
+                        </div>
                     </div>
                 {{/progress}}
             </div>
         </div>
     {{/milestones}}
 </div>
-<!-- jQuery (necessary for Bootstrap's JavaScript plugins) -->
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
-<!-- Include all compiled plugins (below), or include individual files as needed -->
-<script src="js/bootstrap.min.js"></script>
+<script src="https://code.jquery.com/jquery-1.13.1.slim.min.js"
+        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+        crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+        integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+        crossorigin="anonymous"></script>
+<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+        integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+        crossorigin="anonymous"></script>
 </body>
 </html>
